+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+0.3.0 (2018-05-12)
+------------------
+
+* Disable CSS animation, allow screenshot delay
+* If save_differences==True the diff file will contain an highlighted diff-rect
+
 0.2.3 (2018-04-11)
 ------------------